add computeAncestorWorkspaces
